Exposes the `onEndReachedThreshold` prop via ListView
The existing default value of 40 is not fit for purpose as the RN
FlatList component expects a value between 0 and 1, representing
a distance in the end as a fraction of the visible length of the
list. Defaulting it to a value above 1 causes the list to refresh
infinitely. This change exposes the FlatList prop as a prop of
ListView, allowing it to be set as required.